+++ wsdot/CameraViewController.swift
@@ -47,7 +47,7 @@ class CameraViewController: UIViewController, GADBannerViewDelegate{
         }
         
         // Ad Banner
-        bannerView.adUnitID = ApiKeys.wsdot_ad_string
+        bannerView.adUnitID = ApiKeys.getAdId()
         bannerView.rootViewController = self
         bannerView.load(GADRequest())
         bannerView.delegate = self