+++ wsdot/GoogleAnalytics.swift
@@ -23,7 +23,7 @@ import Foundation
 class GoogleAnalytics {
     
     static let analytics_enabled = true
-    static let analytics_dryrun = true
+    static let analytics_dryrun = false
     
     static func screenView(screenName: String){
         if (GoogleAnalytics.analytics_enabled){