+++ wsdot/HomeSplitViewController.swift
@@ -27,6 +27,10 @@ class HomeSplitViewController: UISplitViewController, UISplitViewControllerDeleg
     override func viewDidLoad() {
         self.delegate = self
         self.preferredDisplayMode = .allVisible
+        
+        let storyboard = UIStoryboard(name: "Favorites", bundle: nil)
+        self.viewControllers.append(contentsOf: [storyboard.instantiateViewController(withIdentifier: "FavoritesNavigationController") as! UINavigationController])
+    
     }
 
     func splitViewController(_ splitViewController: UISplitViewController, collapseSecondary secondaryViewController: UIViewController, onto primaryViewController: UIViewController) -> Bool {