+++ wsdot/TrafficMapViewController.swift
@@ -19,9 +19,9 @@
 //
 
 import UIKit
-import UIKit
 import GoogleMaps
 import GoogleMobileAds
+import EasyTipView
 
 class TrafficMapViewController: UIViewController, MapMarkerDelegate, GMSMapViewDelegate, GMUClusterManagerDelegate, GADBannerViewDelegate {
     
@@ -65,6 +65,9 @@ class TrafficMapViewController: UIViewController, MapMarkerDelegate, GMSMapViewD
     
     fileprivate let closedIconImage = UIImage(named: "icMapClosed")
     
+    var tipView = EasyTipView(text: "")
+    
+    @IBOutlet weak var toolBarMenuButton: UIBarButtonItem!
     @IBOutlet weak var cameraBarButton: UIBarButtonItem!
     @IBOutlet weak var bannerView: GADBannerView!
     @IBOutlet weak var activityIndicatorView: UIActivityIndicatorView!
@@ -95,7 +98,7 @@ class TrafficMapViewController: UIViewController, MapMarkerDelegate, GMSMapViewD
         embeddedMapViewController.clusterManager.setDelegate(self, mapDelegate: self)
         
         // Ad Banner
-        bannerView.adUnitID = ApiKeys.wsdot_ad_string
+        bannerView.adUnitID = ApiKeys.getAdId()
         bannerView.rootViewController = self
         bannerView.load(GADRequest())
         bannerView.delegate = self
@@ -111,6 +114,7 @@ class TrafficMapViewController: UIViewController, MapMarkerDelegate, GMSMapViewD
         super.viewWillAppear(animated)
         GoogleAnalytics.screenView(screenName: "/Traffic Map")
     }
+
     
     @IBAction func refreshPressed(_ sender: UIBarButtonItem) {
     
@@ -164,6 +168,7 @@ class TrafficMapViewController: UIViewController, MapMarkerDelegate, GMSMapViewD
     }
 
     @IBAction func settingsAction(_ sender: UIBarButtonItem) {
+        tipView.dismiss()
         performSegue(withIdentifier: SegueSettingsPopover, sender: self)
     }
     
@@ -689,4 +694,24 @@ class TrafficMapViewController: UIViewController, MapMarkerDelegate, GMSMapViewD
         }
     }
 
+}
+
+extension TrafficMapViewController: EasyTipViewDelegate {
+    
+    public func easyTipViewDidDismiss(_ tipView: EasyTipView) {
+         UserDefaults.standard.set(true, forKey: UserDefaultsKeys.hasSeenClusterTipView)
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        tipView.dismiss()
+    }
+
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        if (!UserDefaults.standard.bool(forKey: UserDefaultsKeys.hasSeenClusterTipView) && !UIAccessibilityIsVoiceOverRunning()){
+            tipView = EasyTipView(text: "Make the map eaiser to read in areas with a high density of cameras by turning on Camera Clustering.", delegate: self)
+            tipView.show(forItem: self.toolBarMenuButton)
+        }
+    }
 }