+++ CHANGELOG.md
@@ -4,7 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: 
 
-## [5.1.0] 2017-01-
+## [5.2.0] 2017-03
+
+### Added
+* My Routes section. Users can record routes and let the WSDOT app add content on their route to their favorites, as well as check for traffic alerts on thir route.
+* Favorites Settings. From this menu users can rearrage their favorites and clear their favorites list.
+* Renaming favorite locations.
+
+## [5.1.0] 2017-01-30
 ### Added
 * When tapping overlapping traffic alerts, a list of all the overlapping alerts will open.
 * WSDOT East Twitter account.